Firestore config
