+++ README.md
@@ -7,5 +7,10 @@ It only runs with live server
 
 1. Basic config
 
-- By creating a basic HTML form, at the end of the body tag we'll add type="module" so we can import the functions of firebase.
-- index.js file will have the basic config from the firebase documentation
+ - By creating a basic HTML form, at the end of the body tag we'll add type="module" so we can use it as a module where we export and import functions.
+ - firebase.js file will have the basic config from the firebase documentation
+
+2. Firestore config
+
+ - At index.js we'll start by adding an eventListener that will take the values from title and description and use the function saveTask imported from our firebase.js
+ - Firebase.js has the basic configuration of our database, which was imported as CDN
